fix: Respect hidden statusline for size calculation
The window_height calculation hardcoded subtracting 1 for the height of
the statusline.
But this is not correct, if the statusline is not visible, which is the
case if either the user sets laststatus=0 or sets laststatus=1 and has
more than one window open.

Fix this by respecting the laststatus setting and adapt the calculation
accordingly.
This is pretty much identical to the tabline height calculation code
that already existed below.